Fix pylint issues like duplicate code
Refactor code to remove duplicate functions and improve maintainability.

* **Add `app/utils.py`**:
  - Add utility functions: `get_token`, `extract_text_from_docx`, `send_text_to_openai`, and `create_response_format`.

* **Modify `app/routes/minnisblad.py`**:
  - Import utility functions from `app/utils.py`.
  - Remove duplicate code for `get_token`, `extract_text_from_docx`, `send_text_to_openai`, and `create_response_format`.

* **Modify `app/routes/minnisblad_adstod.py`**:
  - Import utility functions from `app/utils.py`.
  - Remove duplicate code for `get_token`, `extract_text_from_docx`, `send_text_to_openai`, and `create_response_format`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/StefnirKristjansson/frodi?shareId=XXXX-XXXX-XXXX-XXXX).
