+++ app/routes/minnisblad.py
@@ -12,15 +12,10 @@ from fastapi import (
     HTTPException,
     Form,
     Depends,
-    status,
 )
 from fastapi.responses import FileResponse, HTMLResponse
-from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from fastapi.templating import Jinja2Templates
-from docx import Document
-from dotenv import load_dotenv
-from openai import OpenAI
-
+from app.utils import get_token, extract_text_from_docx, send_text_to_openai, create_response_format
 
 templates = Jinja2Templates(directory="app/templates")
 
@@ -28,33 +23,6 @@ router = APIRouter()
 
 load_dotenv()
 BEARER_TOKEN = os.getenv("BEARER_TOKEN")
-
-# Security scheme and token validation
-token_auth_scheme = HTTPBearer()
-
-
-def get_token(credentials: HTTPAuthorizationCredentials = Depends(token_auth_scheme)):
-    """Validate the token and return it if it is valid."""
-    if credentials.scheme != "Bearer":
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid authentication scheme.",
-        )
-    if credentials.credentials != BEARER_TOKEN:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid or expired token.",
-        )
-    return credentials.credentials
-
-
-openai_api_key = os.getenv("OPENAI_API_KEY")
-
-# Set the OpenAI API key globally
-
-OpenAI.api_key = openai_api_key
-client = OpenAI()
-
 
 @router.get("/minnisblad", response_class=HTMLResponse)
 async def minnisblad(request: Request):
@@ -110,131 +78,6 @@ async def upload_file(
         return templates.TemplateResponse(
             "index.html", {"request": request, "error": str(e)}
         )
-
-
-def extract_text_from_docx(file):
-    """Extract text from a .docx file."""
-    doc = Document(file)
-    text = []
-    for paragraph in doc.paragraphs:
-        text.append(paragraph.text)
-    return "\n".join(text)
-
-
-async def send_text_to_openai(
-    text: str, response_format: dict
-) -> dict:  # pragma: no cover
-    """Send the text to the OpenAI API and return the response."""
-    # Using the OpenAI client as per your original
-    content_text = """Notandinn sendir þér texta sem þú átt að breyta í minnisblað.
-    Bættu í og styttu textan eftir þörfum og kaflaskiftu eins og þú sérð best. 
-    Passaðu að hafa kaflanna ekki of stutta ekki hafa fleirri en 2 kafla á blaðsíðu."""
-    messages = [
-        {
-            "role": "system",
-            "content": [
-                {
-                    "type": "text",
-                    "text": content_text,
-                }
-            ],
-        },
-        {
-            "role": "user",
-            "content": [{"type": "text", "text": text}],
-        },
-    ]
-
-    completion = client.chat.completions.create(
-        model="gpt-4o",
-        messages=messages,
-        temperature=1,
-        max_tokens=2048,
-        top_p=1,
-        frequency_penalty=0,
-        presence_penalty=0,
-        response_format=response_format,
-    )
-    response = json.loads(completion.choices[0].message.content)
-    return response
-
-
-def create_response_format(selected_chapters: list) -> dict:
-    """Create the response format based on the selected chapters."""
-    base_response = {
-        "type": "json_schema",
-        "json_schema": {
-            "name": "minnisblad",  # Updated name
-            "strict": True,
-            "schema": {
-                "type": "object",
-                "properties": {
-                    "titill": {
-                        "type": "string",
-                        "description": "Titill minnisblaðsins.",
-                    },
-                    "kaflar": {
-                        "type": "array",
-                        "description": "Kaflar minnisblaðsins.",
-                        "items": {
-                            "type": "object",
-                            "properties": {
-                                "chapter_title": {
-                                    "type": "string",
-                                    "description": "Titill kafla.",
-                                },
-                                "content": {
-                                    "type": "string",
-                                    "description": "Innihald kafla.",
-                                },
-                            },
-                            "required": ["chapter_title", "content"],
-                            "additionalProperties": False,
-                        },
-                    },
-                },
-                "required": [
-                    "titill",
-                    "kaflar",
-                ],
-                "additionalProperties": False,
-            },
-        },
-    }
-    inngangur_description = "Inngangur minnisblaðsins."
-    samantekt_description = "Samantekt minnisblaðsins."
-    aaetlun_description = "Áætlun minnisblaðsins."
-    markmid_description = "Markmið minnisblaðsins."
-
-    if "inngangur" in selected_chapters:
-        base_response["json_schema"]["schema"]["properties"]["Inngangur"] = {
-            "type": "string",
-            "description": inngangur_description,
-        }
-        base_response["json_schema"]["schema"]["required"].append("Inngangur")
-
-    if "samantekt" in selected_chapters:
-        base_response["json_schema"]["schema"]["properties"]["Samantekt"] = {
-            "type": "string",
-            "description": samantekt_description,
-        }
-        base_response["json_schema"]["schema"]["required"].append("Samantekt")
-
-    if "aaetlun" in selected_chapters:
-        base_response["json_schema"]["schema"]["properties"]["Áætlun"] = {
-            "type": "string",
-            "description": aaetlun_description,
-        }
-        base_response["json_schema"]["schema"]["required"].append("Áætlun")
-
-    if "markmid" in selected_chapters:
-        base_response["json_schema"]["schema"]["properties"]["Markmið"] = {
-            "type": "string",
-            "description": markmid_description,
-        }
-        base_response["json_schema"]["schema"]["required"].append("Markmið")
-
-    return base_response
 
 
 def create_docx_from_json(response_json: dict) -> str:
