+++ app/routes/minnisblad_adstod.py
@@ -13,12 +13,8 @@ from fastapi import (
     status,
 )
 from fastapi.responses import JSONResponse, HTMLResponse
-from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from fastapi.templating import Jinja2Templates
-from docx import Document
-from dotenv import load_dotenv
-from openai import OpenAI
-
+from app.utils import get_token, extract_text_from_docx, send_text_to_openai, create_response_format
 
 templates = Jinja2Templates(directory="app/templates")
 
@@ -26,25 +22,6 @@ router = APIRouter()
 
 load_dotenv()
 BEARER_TOKEN = os.getenv("BEARER_TOKEN")
-
-# Security scheme and token validation
-token_auth_scheme = HTTPBearer()
-
-
-def get_token(credentials: HTTPAuthorizationCredentials = Depends(token_auth_scheme)):
-    """Validate the token and return it if it is valid."""
-    if credentials.scheme != "Bearer":
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid authentication scheme.",
-        )
-    if credentials.credentials != BEARER_TOKEN:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid or expired token.",
-        )
-    return credentials.credentials
-
 
 openai_api_key = os.getenv("OPENAI_API_KEY")
 
@@ -97,102 +74,3 @@ async def upload_file(
         return templates.TemplateResponse(
             "index.html", {"request": request, "error": str(e)}
         )
-
-
-def extract_text_from_docx(file):
-    """Extract text from a .docx file."""
-    doc = Document(file)
-    text = []
-    for paragraph in doc.paragraphs:
-        text.append(paragraph.text)
-    return "\n".join(text)
-
-
-async def send_text_to_openai(
-    text: str, response_format: dict
-) -> dict:  # pragma: no cover
-    """Send the text to the OpenAI API and return the response."""
-    # Using the OpenAI client as per your original
-    content_text = """Notendinn sendi þér minnisblað, farðu mjög varlega yfir það og 
-    finndu dæmi um önnur minnisblöð, 
-    gefðu þér tíma að skoa Íslenskt málfar og stafsetningu."""
-    messages = [
-        {
-            "role": "system",
-            "content": [
-                {
-                    "type": "text",
-                    "text": content_text,
-                }
-            ],
-        },
-        {
-            "role": "user",
-            "content": [{"type": "text", "text": text}],
-        },
-    ]
-
-    completion = client.chat.completions.create(
-        model="gpt-4o",
-        messages=messages,
-        temperature=1,
-        max_tokens=4000,
-        top_p=1,
-        frequency_penalty=0,
-        presence_penalty=0,
-        response_format=response_format,
-    )
-    response = json.loads(completion.choices[0].message.content)
-    return response
-
-
-def create_response_format(selected_chapters: list) -> dict:
-    """Create the response format based on the selected chapters."""
-    base_response = base_response = {
-        "type": "json_schema",
-        "json_schema": {
-            "name": "Minnisblad_adstod",
-            "schema": {  # Correcting to include 'schema' directly
-                "title": "Minnisblad_adstod",
-                "type": "object",
-                "properties": {
-                    "name": {"type": "string", "description": "The name of the schema"},
-                    "type": {
-                        "type": "string",
-                        "enum": [
-                            "object",
-                            "array",
-                            "string",
-                            "number",
-                            "boolean",
-                            "null",
-                        ],
-                        "description": "The type of the schema (should match JSON Schema types)",
-                    },
-                    "properties": {
-                        "type": "object",
-                        "properties": {
-                            "malfar": {
-                                "type": "string",
-                                "description": "Hérna setur þú inn hvað meigi bæta og breyta varðandi málfar, stíl og uppbyggingu minnisblaðsins.",
-                            },
-                            "stafsetning": {
-                                "type": "string",
-                                "description": "Hérna setur þú inn hvað meigi bæta og breyta varðandi stafsetningu passaðu að benda skýrt á allar stafsetningarvillur með að vitna í textan sem á við.",
-                            },
-                            "radleggingar": {
-                                "type": "string",
-                                "description": "Hérna setur þú inn hvað meigi bæta og breyta varðandi almennar raðleggingar. Farðu vel yfir hvernig minnisblöð eru yfir höfuð og hafðu það í huga",
-                            },
-                        },
-                        "required": ["malfar", "stafsetning", "radleggingar"],
-                        "additionalProperties": False,
-                    },
-                },
-                "required": ["name", "type", "properties"],
-                "additionalProperties": False,
-            },
-        },
-    }
-
-    return base_response
